# firewall/openapi/telco-openapi-one.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Insecure Telco Vulnerable API",
    "version": "1.0.0",
    "description": "An intentionally vulnerable API inspired by TMF 622, designed to eventually showcase API security issues like BOLA,BOPLA, BFLA, and other OWASP API Top 10 vulnerabilities."
  },
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "User login",
        "operationId": "UserLogin",
        "description": "Authenticates a user with a username and password and returns a JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "$ref": "#/components/schemas/username"
                  },
                  "password": {
                    "$ref": "#/components/schemas/password"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "$ref": "#/components/schemas/username"
                        },
                        "password": {
                          "$ref": "#/components/schemas/password"
                        },
                        "email": {
                          "$ref": "#/components/schemas/email"
                        },
                        "isAdmin": {
                          "type": "boolean"
                        },
                        "customerId": {
                          "$ref": "#/components/schemas/customerId"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ResponseError400"
          },
          "401": {
            "$ref": "#/components/responses/ResponseError401"
          },
          "406": {
            "$ref": "#/components/responses/ResponseError406"
          },
          "415": {
            "$ref": "#/components/responses/ResponseError415"
          },
          "429": {
            "$ref": "#/components/responses/ResponseError429"
          },
          "500": {
            "$ref": "#/components/responses/ResponseError500"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "User registration",
        "operationId": "UserRegistration",
        "description": "Registers a new user with a username, password, and email.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "$ref": "#/components/schemas/username"
                  },
                  "password": {
                    "$ref": "#/components/schemas/password"
                  },
                  "email": {
                    "$ref": "#/components/schemas/email"
                  },
                  "address": {
                    "$ref": "#/components/schemas/address"
                  }
                },
                "required": [
                  "username",
                  "password",
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User registered successfully",
                      "pattern": "^User registered successfully$",
                      "minLength": 0,
                      "maxLength": 28
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "username": {
                          "$ref": "#/components/schemas/username"
                        },
                        "email": {
                          "$ref": "#/components/schemas/email"
                        },
                        "isAdmin": {
                          "type": "boolean",
                          "example": false
                        },
                        "customerId": {
                          "$ref": "#/components/schemas/customerId"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ResponseError400"
          },
          "401": {
            "$ref": "#/components/responses/ResponseError401"
          },
          "406": {
            "$ref": "#/components/responses/ResponseError406"
          },
          "415": {
            "$ref": "#/components/responses/ResponseError415"
          },
          "429": {
            "$ref": "#/components/responses/ResponseError429"
          },
          "500": {
            "$ref": "#/components/responses/ResponseError500"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      }
    },
    "/auth/user/{username}": {
      "delete": {
        "summary": "Delete User",
        "operationId": "DeleteUser",
        "description": "Deletes a user by their username. Only the user themselves or an admin can perform this action.",
        "parameters": [
          {
            "$ref": "#/components/parameters/username"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ResponseError401"
          },
          "400": {
            "$ref": "#/components/responses/ResponseError400"
          },
          "406": {
            "$ref": "#/components/responses/ResponseError406"
          },
          "415": {
            "$ref": "#/components/responses/ResponseError415"
          },
          "429": {
            "$ref": "#/components/responses/ResponseError429"
          },
          "403": {
            "$ref": "#/components/responses/ResponseError403"
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ResponseError500"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "username": {
        "name": "username",
        "in": "path",
        "description": "The unique username of the user to delete.",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "username": {
        "type": "string"
      },
      "password": {
        "type": "string"
      },
      "email": {
        "type": "string"
      },
      "address": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
...
      "customerId": {
        "type": "string"
      },
...
      "MessageObject": {
        "type": "object",
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ResponseMessage"
          }
        }
      },
      "ResponseMessage": {
        "type": "string"
      }
    },
    "responses": {
      "DefaultErrorResponse": {
        "description": "Unexpected Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError400": {
        "description": "Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
...
      "ResponseError406": {
        "description": "Not Acceptable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError415": {
        "description": "Unsupported Media Type",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError429": {
        "description": "Too Many Requests",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      },
      "ResponseError500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageObject"
            }
          }
        }
      }
    }
  }
}